Refactored method into common file
