+++ main/integration-tests/__test__/api-tests/workspace-service-catalogs/connections/windows-rdp-info-connection.test.js
@@ -16,6 +16,10 @@
 const { runSetup } = require('../../../../support/setup');
 const errorCode = require('../../../../support/utils/error-code');
 
+const {
+  createWorkspaceTypeAndConfiguration,
+} = require('../../../../support/complex/create-workspace-type-and-configuration');
+
 describe('Get Windows password for RDP scenario', () => {
   let setup;
   let adminSession;
@@ -29,24 +33,11 @@ describe('Get Windows password for RDP scenario', () => {
     await setup.cleanup();
   });
 
-  async function createWorkspace(allowRoleIds = ['admin']) {
-    const workspaceTypeId = setup.gen.string({ prefix: 'workspace-test' });
-    const configurationId = setup.gen.string({ prefix: 'configuration-test' });
-
-    await adminSession.resources.workspaceTypes.create({ id: workspaceTypeId, status: 'approved' });
-    await adminSession.resources.workspaceTypes
-      .workspaceType(workspaceTypeId)
-      .configurations()
-      .create({ id: configurationId, allowRoleIds });
-
-    return { workspaceTypeId, configurationId };
-  }
-
   describe('Get Windows password for RDP scenario', () => {
     it('should fail if user is inactive', async () => {
       const adminSession2 = await setup.createAdminSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
       const connectionId = setup.gen.string({ prefix: 'workspace-service-catalog-connection-test' });
 
       const response = await adminSession.resources.workspaceServiceCatalogs.create({
@@ -71,7 +62,7 @@ describe('Get Windows password for RDP scenario', () => {
     it('should fail if user is anonymous', async () => {
       const anonymousSession = await setup.createAnonymousSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
       const connectionId = setup.gen.string({ prefix: 'workspace-service-catalog-connection-test' });
 
       const response = await adminSession.resources.workspaceServiceCatalogs.create({
