+++ main/integration-tests/__test__/api-tests/workspace-service-catalogs/create-workspace-service-catalog.test.js
@@ -14,6 +14,10 @@
  */
 
 const { runSetup } = require('../../../support/setup');
+
+const {
+  createWorkspaceTypeAndConfiguration,
+} = require('../../../support/complex/create-workspace-type-and-configuration');
 const errorCode = require('../../../support/utils/error-code');
 
 describe('Create workspace-service-catalog scenarios', () => {
@@ -29,24 +33,11 @@ describe('Create workspace-service-catalog scenarios', () => {
     await setup.cleanup();
   });
 
-  async function createWorkspace(allowRoleIds = ['admin']) {
-    const workspaceTypeId = setup.gen.string({ prefix: 'workspace-test' });
-    const configurationId = setup.gen.string({ prefix: 'configuration-test' });
-
-    await adminSession.resources.workspaceTypes.create({ id: workspaceTypeId, status: 'approved' });
-    await adminSession.resources.workspaceTypes
-      .workspaceType(workspaceTypeId)
-      .configurations()
-      .create({ id: configurationId, allowRoleIds });
-
-    return { workspaceTypeId, configurationId };
-  }
-
   describe('Create workspace-service-catalog', () => {
     it('should fail if user is inactive', async () => {
       const adminSession2 = await setup.createAdminSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await adminSession2.resources.users.deactivateUser(adminSession2.user);
 
@@ -64,7 +55,7 @@ describe('Create workspace-service-catalog scenarios', () => {
     it('should fail if user is anonymous', async () => {
       const anonymousSession = await setup.createAnonymousSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await expect(
         anonymousSession.resources.workspaceServiceCatalogs.create({
@@ -80,7 +71,7 @@ describe('Create workspace-service-catalog scenarios', () => {
     it('should fail if user role is not allowed', async () => {
       const researcherSession = await setup.createResearcherSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await expect(
         researcherSession.resources.workspaceServiceCatalogs.create({
@@ -96,7 +87,7 @@ describe('Create workspace-service-catalog scenarios', () => {
 
     it('should fail if input is not valid', async () => {
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await expect(
         adminSession.resources.workspaceServiceCatalogs.create({
@@ -112,7 +103,7 @@ describe('Create workspace-service-catalog scenarios', () => {
 
     it('should create the service catalog workspace if admin', async () => {
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await expect(
         adminSession.resources.workspaceServiceCatalogs.create({
@@ -129,7 +120,9 @@ describe('Create workspace-service-catalog scenarios', () => {
     it('should create if user role is allowed', async () => {
       const researcherSession = await setup.createResearcherSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace(['researcher']);
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup, [
+        'researcher',
+      ]);
 
       await expect(
         researcherSession.resources.workspaceServiceCatalogs.create({
