+++ main/integration-tests/__test__/api-tests/workspace-service-catalogs/get-workspace-service-catalog.test.js
@@ -14,6 +14,10 @@
  */
 
 const { runSetup } = require('../../../support/setup');
+
+const {
+  createWorkspaceTypeAndConfiguration,
+} = require('../../../support/complex/create-workspace-type-and-configuration');
 const errorCode = require('../../../support/utils/error-code');
 
 describe('Get workspace-service-catalog scenarios', () => {
@@ -29,24 +33,11 @@ describe('Get workspace-service-catalog scenarios', () => {
     await setup.cleanup();
   });
 
-  async function createWorkspace(allowRoleIds = ['admin']) {
-    const workspaceTypeId = setup.gen.string({ prefix: 'workspace-test' });
-    const configurationId = setup.gen.string({ prefix: 'configuration-test' });
-
-    await adminSession.resources.workspaceTypes.create({ id: workspaceTypeId, status: 'approved' });
-    await adminSession.resources.workspaceTypes
-      .workspaceType(workspaceTypeId)
-      .configurations()
-      .create({ id: configurationId, allowRoleIds });
-
-    return { workspaceTypeId, configurationId };
-  }
-
   describe('Get workspace-service-catalog', () => {
     it('should fail if user is inactive', async () => {
       const adminSession2 = await setup.createAdminSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       await adminSession.resources.users.deactivateUser(adminSession2.user);
 
@@ -66,7 +57,7 @@ describe('Get workspace-service-catalog scenarios', () => {
     it('should fail if user is anonymous', async () => {
       const anonymousSession = await setup.createAnonymousSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       const response = await adminSession.resources.workspaceServiceCatalogs.create({
         name: workspaceName,
@@ -84,7 +75,7 @@ describe('Get workspace-service-catalog scenarios', () => {
     it('should fail if user role is not owner', async () => {
       const researcherSession = await setup.createResearcherSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       const response = await adminSession.resources.workspaceServiceCatalogs.create({
         name: workspaceName,
@@ -102,7 +93,9 @@ describe('Get workspace-service-catalog scenarios', () => {
     it('should pass if user role is owner', async () => {
       const researcherSession = await setup.createResearcherSession();
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace(['researcher']);
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup, [
+        'researcher',
+      ]);
 
       const response = await researcherSession.resources.workspaceServiceCatalogs.create({
         name: workspaceName,
@@ -120,7 +113,7 @@ describe('Get workspace-service-catalog scenarios', () => {
 
     it('should pass if user is admin', async () => {
       const workspaceName = setup.gen.string({ prefix: 'workspace-service-catalog-test' });
-      const { workspaceTypeId, configurationId } = await createWorkspace();
+      const { workspaceTypeId, configurationId } = await createWorkspaceTypeAndConfiguration(adminSession, setup);
 
       const response = await adminSession.resources.workspaceServiceCatalogs.create({
         name: workspaceName,
