+++ main/integration-tests/support/complex/create-workspace-type-and-configuration.js
@@ -1,0 +1,22 @@
+/**
+ * A function that performs the complex task of creating a workspace type
+ * and configuration.
+ *
+ * @param {string} adminSession The admin session.
+ * @param {string} setup The setup object.
+ * @param {Array} allowRoledIds The roles allowed to launch the workspace.
+ */
+async function createWorkspaceTypeAndConfiguration(adminSession, setup, allowRoleIds = ['admin']) {
+  const workspaceTypeId = setup.gen.string({ prefix: 'workspace-test' });
+  const configurationId = setup.gen.string({ prefix: 'configuration-test' });
+
+  await adminSession.resources.workspaceTypes.create({ id: workspaceTypeId, status: 'approved' });
+  await adminSession.resources.workspaceTypes
+    .workspaceType(workspaceTypeId)
+    .configurations()
+    .create({ id: configurationId, allowRoleIds });
+
+  return { workspaceTypeId, configurationId };
+}
+
+module.exports = { createWorkspaceTypeAndConfiguration };
